add ncncjd
