+++ .idea/libraries/Maven__io_pivotal_gemfire_geode_wan_9_1_1.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: io.pivotal.gemfire:geode-wan:9.1.1">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/io/pivotal/gemfire/geode-wan/9.1.1/geode-wan-9.1.1.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/io/pivotal/gemfire/geode-wan/9.1.1/geode-wan-9.1.1-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/io/pivotal/gemfire/geode-wan/9.1.1/geode-wan-9.1.1-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
