+++ .idea/libraries/Maven__org_gj_demo_gemfire_common_0_0_1_SNAPSHOT.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: org.gj.demo:gemfire-common:0.0.1-SNAPSHOT">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/gj/demo/gemfire-common/0.0.1-SNAPSHOT/gemfire-common-0.0.1-SNAPSHOT.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/org/gj/demo/gemfire-common/0.0.1-SNAPSHOT/gemfire-common-0.0.1-SNAPSHOT-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/gj/demo/gemfire-common/0.0.1-SNAPSHOT/gemfire-common-0.0.1-SNAPSHOT-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
