+++ .idea/libraries/Maven__org_springframework_data_spring_data_gemfire_2_0_1_RELEASE.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: org.springframework.data:spring-data-gemfire:2.0.1.RELEASE">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/springframework/data/spring-data-gemfire/2.0.1.RELEASE/spring-data-gemfire-2.0.1.RELEASE.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/org/springframework/data/spring-data-gemfire/2.0.1.RELEASE/spring-data-gemfire-2.0.1.RELEASE-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/springframework/data/spring-data-gemfire/2.0.1.RELEASE/spring-data-gemfire-2.0.1.RELEASE-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
